% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate_matrix.R
\name{animate_matrix}
\alias{animate_matrix}
\title{Animate matrix}
\usage{
animate_matrix(
  transformations,
  points_start = NULL,
  return_graph_objects = FALSE,
  return_static = FALSE,
  datasaurus = FALSE
)
}
\arguments{
\item{transformations}{List of matrix transformations to apply (length will
determine total number of explicit frames in animation.)}

\item{points_start}{Dataframe containing x and y columns of points that will
be plotted (if any coordinate is greater than magnitude 5, coordinates will
be automatically scaled down).}

\item{return_graph_objects}{Logical indicating whether to return the
dataframe containing the objects that would have gone into the animation.}

\item{return_static}{Logical indicating whether to return the static ggplot
graph.}

\item{datasaurus}{Logical, where \code{TRUE} is equivalent to is equivalent to
setting \code{points_start = filter(datasauRus::datasaurus_dozen, dataset == "dino")}}
}
\value{
An animation created by \code{gganimate}.
}
\description{
Animate matrix
}
