% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animatrixr-package.R
\docType{package}
\name{animatrixr}
\alias{animatrixr}
\alias{_PACKAGE}
\alias{animatrixr-package}
\title{Animate 2x2 transformation matrices}
\description{
Toy package in support of my blog post on visualizing simple 2x2
matrix transformations. Not meant for general use.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jtipton25/dasc2594}
  \item Report bugs at \url{https://github.com/jtipton25/dasc2594/issues}
}

}
\author{
\strong{Maintainer}: John Tipton \email{jrtipton@uark.edu}

}
\keyword{internal}
