% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_to_system.R
\name{array_to_system}
\alias{array_to_system}
\title{Convert an R array into a \LaTeX system of equations for easy writing in Rmarkdown}
\usage{
array_to_system(A)
}
\arguments{
\item{A}{An array to convert to Latex format}
}
\value{
A string that can be put into an Rmarkdown document by using \code{cat()} to direct the output
}
\description{
Convert an R array into a \LaTeX system of equations for easy writing in Rmarkdown
}
\examples{
latex_string <- array_to_latex(matrix(1:6, 3, 2))

# within Rmarkdown chunk
# cat(latex_string)
}
