% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementary_matrix.R
\name{elementary_matrix_sequence}
\alias{elementary_matrix_sequence}
\title{Calculate a random elementary matrix}
\usage{
elementary_matrix_sequence(n, k, verbose = FALSE)
}
\arguments{
\item{n}{The size of the elementary matrix (an n times n matrix)}

\item{k}{The number of elemenatry row operations to perform}

\item{verbose}{A logical value that determines whether to print the output}
}
\value{
An product of elementary matrices, the inverse of the produce of elementary matrices, and the full sequence of elementary matrices
}
\description{
Calculate a random elementary matrix
}
