% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_basis.R
\name{is_basis}
\alias{is_basis}
\title{Checks if the columns of a square matrix define a basis}
\usage{
is_basis(B)
}
\arguments{
\item{B}{A \eqn{n \times n}{n by n} matrix}
}
\value{
A logical of whether the matrix is a basis
}
\description{
Checks if the columns of a square matrix define a basis
}
