% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_consistent.R
\name{is_consistent}
\alias{is_consistent}
\title{Check if the matrix equation Ax = b is consistent}
\usage{
is_consistent(A, b)
}
\arguments{
\item{A}{The matrix A}

\item{b}{The vector b}
}
\value{
This function returns TRUE if there is a solution to the equation Ax = b and FALSE if there is not a solution to the equation Ax = b
}
\description{
Check if the matrix equation Ax = b is consistent
}
\examples{

library(dasc2594)
A <- diag(2)
b <- 1:2
# expect TRUE
is_consistent(A, b)

A <- matrix(c(1, 0, 1, 0), 2, 2)
b <- 1:2
# expect FALSE
is_consistent(A, b)
}
