% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_invertible.R
\name{is_invertible}
\alias{is_invertible}
\title{Test if a matrix A is invertible}
\usage{
is_invertible(A)
}
\arguments{
\item{A}{a matrix}
}
\value{
TRUE if A is invertible and FALSE if A is not invertible
}
\description{
Test if a matrix A is invertible
}
\examples{

library(dasc2594)
# An invertible matrix
A <- diag(4)
is_invertible(A) # returns TRUE

# A non-invertible matrix
A <- matrix(1, 3, 3)
is_invertible(A) # returns FALSE
}
