% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_unique.R
\name{is_unique}
\alias{is_unique}
\title{Check if the matrix equation Ax = b has a unique solution}
\usage{
is_unique(A, b)
}
\arguments{
\item{A}{The matrix A}

\item{b}{The vector b}
}
\value{
This function returns TRUE if there is a unique solution to the equation Ax = b and FALSE if there is not a unique solution to the equation Ax = b. If there is no solution to the system of equations, the function will return an error.


}
\description{
Check if the matrix equation Ax = b has a unique solution
}
\examples{

library(dasc2594)
A <- diag(2)
b <- 1:2
# expect TRUE
is_consistent(A, b)
# expect TRUE
is_unique(A, b)

A <- matrix(c(1, 0, 0, 1, 1, 1), 2, 3)
b <- 1:2
# expect FALSE
is_unique(A, b)
}
