% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_operations.R
\name{is_valid_row}
\alias{is_valid_row}
\title{Check if the row is valid for the matrix A}
\usage{
is_valid_row(A, row)
}
\arguments{
\item{A}{The matrix A}

\item{row}{The row to check if it is valid}
}
\value{
\code{is_valid_row()} returns TRUE if \code{row} is a valid row for \code{A} and \code{FALSE} otherwise
}
\description{
Check if the row is valid for the matrix A
}
\examples{
A <- diag(4)
# expect TRUE
is_valid_row(A, 3)
# expect FALSE
is_valid_row(A, 5)

}
