% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_vector}
\alias{is_vector}
\title{Check if input is a vector or a matrix with one column}
\usage{
is_vector(x)
}
\arguments{
\item{x}{is the input}
}
\description{
this function checks if the input is aa matrix with one column
}
\keyword{internal}
