% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex_equation_to_png.R
\name{latex_equation_to_png}
\alias{latex_equation_to_png}
\title{Write an equation in Rmarkdown to a LaTeX format for compiling into a png equation image}
\usage{
latex_equation_to_png(A, b, file = NULL, format = "system of equations")
}
\arguments{
\item{A}{An array to convert to Latex format}

\item{b}{A vector to convert to Latex format}

\item{file}{The file location to save the pdf output}
}
\value{
A string that can be put into an Rmarkdown document by using \code{cat()} to direct the output
}
\description{
Write an equation in Rmarkdown to a LaTeX format for compiling into a png equation image
}
\examples{
latex_string <- array_to_latex(matrix(1:6, 3, 2))

# within Rmarkdown chunk
# cat(latex_string)
}
