% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_basis.R
\name{make_basis}
\alias{make_basis}
\title{Generate a random basis for R^n}
\usage{
make_basis(n)
}
\arguments{
\item{n}{The dimension of the basis generated}
}
\value{
A n by n matrix where each column is a basis vector
}
\description{
Generate a random basis for R^n
}
\examples{
library(dasc2594)
B <- make_basis(4)
# check that columns of B are linearly independent
all.equal(rref(B), diag(4))
}
