% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_eigen.R
\name{make_eigen}
\alias{make_eigen}
\title{Make a matrix with integer eigen factorizations}
\usage{
make_eigen(n)
}
\arguments{
\item{n}{The dimension of the matrix}
}
\value{
A list with three components. A is the matrix that has eigen
decomposition A = V D V^{-1} where V is a matrix with eigenvectors
as columns and D = diag(d) is a matrix with corresponding eigenvalues d
on the diagonal.
}
\description{
Make a matrix with integer eigen factorizations
}
\examples{

e <- make_eigen(3)
e$A
e$V
e$d
}
