% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_system_of_equations.R
\name{make_system_of_equations}
\alias{make_system_of_equations}
\title{Simulate a system of equations Ax = b
Simulate a systemt of equations Ax = b that is (usually) easy to solve using RREF without needing many complicated fractions.}
\usage{
make_system_of_equations(
  n_equations,
  n_variables,
  dim_col = min(n_equations, n_variables),
  dim_null = n_variables - dim_col,
  is_consistent = ifelse(dim_col == n_equations, TRUE, sample(c(TRUE, FALSE), 1)),
  is_homogeneous = FALSE
)
}
\arguments{
\item{n_equations}{The number of equations in the system (the number of rows for the matrix A).}

\item{n_variables}{The number of variables in the system (the number of columns for the matrix A).}

\item{dim_col}{The dimension of the column space. Must be between 1 and the minimum of the number of n_equations and n_variables.}

\item{dim_null}{The dimension of the column space. Must be between 1 and the minimum of the number of n_equations and n_variables.}

\item{is_consistent}{Is the system of equations Ax = b consistent (Is there a solution)?}

\item{is_homogeneous}{Is the vector b equal to the 0 vector?}
}
\value{
A list that contains the matrix A and the vector b. If the system of equations is not homogeneous or consistent, the list will also contain the solution vector x
}
\description{
Simulate a system of equations Ax = b
Simulate a systemt of equations Ax = b that is (usually) easy to solve using RREF without needing many complicated fractions.
}
\examples{
n_equations <- 7
n_variables <- 5
dim_col = min(n_equations, n_variables)
dim_null = n_variables - dim_col
eq <- make_system_of_equations(n_equations, n_variables, dim_col = 2, is_consistent = TRUE)
str(eq)
rref(cbind(eq$A, eq$b))
if(eq$is_consistent)
  all.equal(eq$A \%*\% eq$x, eq$b)
}
