% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tanget_plane.R
\name{plot_tangent_plane}
\alias{plot_tangent_plane}
\title{Plot the function \eqn{f(x, y)} and the tangent plane of \eqn{f(x, y)} at a point \eqn{(a, b)}}
\usage{
plot_tangent_plane(
  target_fun,
  grad_fun,
  a = 3,
  b = 2,
  n = 50,
  xlim = c(-4, 4),
  ylim = c(-4, 4)
)
}
\arguments{
\item{target_fun}{The function \eqn{f(x, y)} of interest.}

\item{grad_fun}{The gradient \eqn{\nabla f(x, y)} of the function \eqn{f(x, y)}.}

\item{a}{The point at which to evaluate the tangent plane in the x-axis coordinates.}

\item{b}{The point at which to evaluate the tangent plane in the y-axis coordinates.}

\item{n}{The number of grid points at which to evaluate the function \eqn{f(x, y)} and the tangent plane of \eqn{f(x, y)}}

\item{xlim}{The range of points in the x-axis at which to evaluate the function \eqn{f(x, y)} and the tangent plane of \eqn{f(x, y)}}

\item{ylim}{The range of points in the y-axis at which to evaluate the function \eqn{f(x, y)} and the tangent plane of \eqn{f(x, y)}}
}
\value{
A plotly plot of the function of interest and the
}
\description{
Plot the function \eqn{f(x, y)} and the tangent plane of \eqn{f(x, y)} at a point \eqn{(a, b)}
}
\examples{
target_fun <- function(x, y) {
    return(x^2 + y^2)
}
  grad_fun <- function(x, y) {
    c(2 * x, 2 * y)
}
plot_tangent_plane(target_fun = target_fun, grad_fun = grad_fun, a=-1, b = 1)
}
