% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_transformation.R
\name{plot_transformation}
\alias{plot_transformation}
\title{plot_transformation}
\usage{
plot_transformation(transformation_matrix)
}
\arguments{
\item{transformation_matrix}{A matrix describing the transformation to animate}
}
\value{
An animation created by \code{gganimate}.
}
\description{
plot_transformation
}
