% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_matrix.R
\name{random_matrix}
\alias{random_matrix}
\title{Generate a random matrix}
\usage{
random_matrix(n, p, values = -9:9)
}
\arguments{
\item{n}{The number of rows}

\item{p}{The number of columns}

\item{values}{A range of values from which to sample with replacement to construct the random matrix}
}
\value{
A n by p matrix with random values sampled with replacement from \code{values}
}
\description{
Generate a random matrix
}
\examples{
library(dasc2594)

# simulate a random vector
u <- random_matrix(3, 1)

# simulate a random matrix
A <- random_matrix(3, 3)

# simulate a random matrix with only positive values
B <- random_matrix(3, 3, values = 1:9)

}
