% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_operations.R
\name{row_add}
\alias{row_add}
\title{Add one row to a multiple of another row}
\usage{
row_add(A, row1, row2, a)
}
\arguments{
\item{A}{The matrix of which to add multiple of one row to another}

\item{row1}{The row to be modified}

\item{row2}{The row to be multiplied and added to row1}

\item{a}{The number to multiply by row 2}
}
\value{
The matrix A where row1 is added to a times row 2
}
\description{
Add one row to a multiple of another row
}
\examples{

A <- diag(3)
row_add(A, 2, 3, -2)

}
