% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_operations.R
\name{row_multiply}
\alias{row_multiply}
\title{Multiply one row by a scalar}
\usage{
row_multiply(A, row, a)
}
\arguments{
\item{A}{The matrix of which to multiply one row}

\item{row}{The row to be modified}

\item{a}{The number to multiply by \code{row}}
}
\value{
The matrix A where row1 is added to a times row 2

A\link{row, } <- a * A\link{row, }
}
\description{
Multiply one row by a scalar
}
\examples{

A <- diag(3)
multiply_row(A, 2, -2)

}
