% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_operations.R
\name{row_swap}
\alias{row_swap}
\title{Swap two rows}
\usage{
row_swap(A, row1, row2)
}
\arguments{
\item{A}{The matrix of which to swap rows}

\item{row1}{The first row to swap}

\item{row2}{The second row to swap}
}
\value{
The matrix A of which to swap rows
}
\description{
Swap two rows
}
\examples{

A <- diag(3)
row_swap(A, 2, 3)

}
