% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tanget_plane.R
\name{tangent_plane}
\alias{tangent_plane}
\title{Evaluate a tangent plane for a function \eqn{f(x, y)} at the point \eqn{(a, b)}.}
\usage{
tangent_plane(x, y, a, b, target_fun, grad_fun)
}
\arguments{
\item{x}{A vector of x-axis coordinates.}

\item{y}{A vector of y-axis coordinates.}

\item{a}{The point at which to evaluate the tangent plane in the x-axis coordinates.}

\item{b}{The point at which to evaluate the tangent plane in the y-axis coordinates.}

\item{target_fun}{The function \eqn{f(x, y)} of interest.}

\item{grad_fun}{The gradient \eqn{\nabla f(x, y)} of the function \eqn{f(x, y)}.}
}
\value{
A vector of values that represent the tangent plane at each input value of \code{x} and \code{y}.
}
\description{
Evaluate a tangent plane for a function \eqn{f(x, y)} at the point \eqn{(a, b)}.
}
