% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_transform_coordinates.R
\name{transform_segment}
\alias{transform_segment}
\title{Transform Coordinates of Segment}
\usage{
transform_segment(df, m)
}
\arguments{
\item{df}{Dataframe with column names \code{x}, \code{y}, \code{xend}, \code{yend}.}

\item{m}{2x2 matrix.}
}
\value{
Dataframe with same column names but coordinates transformed according to \code{m}.
}
\description{
Helper function to apply \code{transform_df_coords()} to the dataframe constructed
by \code{construct_grid()} or similar.
}
