% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_to_latex.R
\name{array_to_latex}
\alias{array_to_latex}
\title{Convert an R array into \LaTeX for easy writing in Rmarkdown}
\usage{
array_to_latex(A, fraction = FALSE)
}
\arguments{
\item{A}{An array to convert to Latex format}

\item{fraction}{An option of whether to print the result as a decimal or fracion}
}
\value{
A string that can be put into an Rmarkdown document by using \code{cat()} to direct the output
}
\description{
Convert an R array into \LaTeX for easy writing in Rmarkdown
}
\examples{
latex_string <- array_to_latex(matrix(1:6, 3, 2))

# within Rmarkdown chunk
# cat(latex_string)
}
