% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementary_matrix.R
\name{elementary_matrix}
\alias{elementary_matrix}
\title{Calculate a random elementary matrix}
\usage{
elementary_matrix(n, operation = NULL, verbose = FALSE)
}
\arguments{
\item{n}{The size of the elementary matrix (an n times n matrix)}

\item{operation}{If specifiec, operation is a character string specifying what kind of elementary row operation to perform. The options are "row swap", "scalar multiplication", and "row addition".}

\item{verbose}{A logical value that determines whether to print the output}
}
\value{
An elementary matrix
}
\description{
Calculate a random elementary matrix
}
