% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_change_basis.R
\name{plot_change_basis}
\alias{plot_change_basis}
\title{plot_change_basis}
\usage{
plot_change_basis(B1, B2)
}
\arguments{
\item{B1}{2 by 2 matrix of the initial basis}

\item{B2}{2 by 2 matrix of the initial basis}
}
\value{
An animation created by \code{gganimate}.
}
\description{
Plot a change of basis from B1 to B2
}
