% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_change_basis.R
\name{plot_change_basis}
\alias{plot_change_basis}
\title{plot_change_basis}
\usage{
plot_change_basis(B1, B2)
}
\arguments{
\item{B1}{2 by 2 matrix of the initial basis}

\item{B2}{2 by 2 matrix of the initial basis}
}
\value{
An animation created by \code{gganimate}.
}
\description{
Plot a change of basis from B1 to B2
}
\examples{

library(tidyverse)
#' if (requireNamespace("patchwork", quietly = TRUE)) {
library(patchwork)
}
# create a 2D basis matrix
B1 <- matrix(c(1, 1, 0, 2), 2, 2)
# create a 2D basis matrix
B2 <- matrix(c(1, 2, 2, 1), 2, 2)

# create a vector with coordinates with respect to the standard basis
x <- c(-1, -2)

# plot the transformation of basis, and the intermediate transformation to the standard basis
p1 <- plot_change_basis(B1, B2) +
  facet_wrap(~ time,
               labeller = labeller(time = c("1" = "B1", "2" = "B2"))) +
  geom_segment(aes(x = 0, xend = -1, y = 0, yend = -2), color = "blue")

 p2 <- plot_change_basis(B1, I) +
   facet_wrap(~ time,
                labeller = labeller(time = c("1" = "B1", "2" = "I"))) +
   geom_segment(aes(x = 0, xend = -1, y = 0, yend = -2), color = "blue")


   p3 <- plot_change_basis(I, B2) +
   facet_wrap(~ time,
              labeller = labeller(time = c("1" = "I", "2" = "B2"))) +
   geom_segment(aes(x = 0, xend = -1, y = 0, yend = -2), color = "blue")

# generate the plot using patchwork
p1 / (p2 + p3)

}
