% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rref.R
\name{rref}
\alias{rref}
\title{rref}
\usage{
rref(A)
}
\arguments{
\item{A}{a matrix}
}
\value{
A matrix in reduced row echelon form
}
\description{
rref
}
\examples{

A <- matrix(c(4, 3, 2, 5, 3, 5, 5, 3, 4, 4, 6, 5), 3, 4)
B <- rref(A)
}
